fix local styled component warnings
