+++ src/components/card/card.jsx
@@ -1,48 +1,17 @@
-import styled from "styled-components";
+import {
+    Container,
+    Image,
+    Video,
+    Details,
+    Owner,
+    TwitterLogo,
+    TwitterHandle
+} from './styles'
 
-const twitterLogo = require("../img/design/twitter.png")
+const twitterLogo = require("../../img/design/twitter.png")
 
 export const Card = ({ nippie }) => {
     const formattedTwitterStr = `https://twitter.com/${nippie.owner.replace('@', '')}`;
-
-    const Container = styled.div`
-        text-align: left;
-        margin-bottom: 30px;
-    `
-
-    const Image = styled.img`
-        outline: 5px solid black;
-        max-width: ${window.innerWidth < 700 ? '150px' : '200px'};
-    `
-
-    const Details = styled.div`
-        margin-left: -5px;
-    `
-
-    const Video = styled.video`
-        outline: 5px solid black;
-        max-width: ${window.innerWidth < 700 ? '150px' : '200px'};
-        width: 200;
-        outline: 10px solid black;
-    `
-
-    const Owner = styled.a`
-        color: black;
-        font-size: 30px;
-        font-weight: bold;
-        text-decoration: none;
-    `
-
-    const TwitterLogo = styled.img`
-        color: black;
-        max-width: 15px;
-        margin-right: 5px;
-    `
-
-    const TwitterHandle = styled.a`
-        color: black;
-        font-size: 18px;
-    `
 
     return (
         <Container>
