--- src/styles.js
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const HiddenAnchor = styled.a`
-    line-height: 0; 
-    font-size: 0;
-    color: transparent; 
-`
